refactor controllers

+++ src/MasterCRM.Application/Services/Clients/ClientService.cs
@@ -1,6 +1,7 @@
 using MasterCRM.Application.MapExtensions;
 using MasterCRM.Application.Services.Clients.Requests;
 using MasterCRM.Application.Services.Clients.Responses;
+using MasterCRM.Domain.Exceptions;
 
 namespace MasterCRM.Application.Services.Clients;
 
@@ -14,33 +15,44 @@ public class ClientService(IClientRepository clientRepository) : IClientService
         return new GetClientsResponse(clientsArray.Length, clientsArray.Select(client => client.ToItemResponse()));
     }
 
-    public async Task<ClientDto?> GetByIdAsync(Guid id)
+    public async Task<ClientDto?> GetByIdAsync(string masterId, Guid id)
     {
         var client = await clientRepository.GetByIdAsync(id);
+
+        if (client == null)
+            return null;
+
+        if (client.MasterId != masterId)
+            throw new ForbidException("Current user is not the owner of the client");
         
-        return client?.ToDto();
+        return client.ToDto();
     }
     
-    public async Task<bool> TryChangeAsync(Guid id, ChangeClientRequest request)
+    public async Task<bool> TryChangeAsync(string masterId, Guid id, ChangeClientRequest request)
     {
         var client = await clientRepository.GetByIdAsync(id);
         
         if (client == null)
             return false;
         
-        //TODO: add history to client's orders
+        if (client.MasterId != masterId)
+            throw new ForbidException("Current user is not the owner of the client");
+        
         client.Update(request.FullName, request.Email, request.Phone);
         
         await clientRepository.SaveChangesAsync();
         return true;
     }
 
-    public async Task<bool> TryDeleteAsync(Guid id)
+    public async Task<bool> TryDeleteAsync(string masterId, Guid id)
     {
         var client = await clientRepository.GetByIdAsync(id);
         
         if (client == null)
             return false;
+        
+        if (client.MasterId != masterId)
+            throw new ForbidException("Current user is not the owner of the client");
         
         clientRepository.Delete(client);
         await clientRepository.SaveChangesAsync();
